mundando o tipo variavel agencia e numero para Long

+++ Conta.java
@@ -1,6 +1,6 @@
 public class Conta {
-    Double agencia;
-    Double numero;
+    Long agencia;
+    Long numero;
     Double saldo;
     String nome;
